Stock_analyze_ai_agent/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
#since we are installing python-dotenc we require a .env file

